Cleaned some casting

+++ Time Left/SKEventCell.m
@@ -34,14 +34,14 @@
         self.deleteButton.hidden = NO;
     
         CABasicAnimation *quiverAnim = [CABasicAnimation animationWithKeyPath:@"transform.rotation"];
-        float startAngle = (-2) * M_PI/180.0;
-        float stopAngle = -startAngle;
+        CGFloat startAngle = (-2) * M_PI/180.0;
+        CGFloat stopAngle = -startAngle;
         quiverAnim.fromValue = [NSNumber numberWithFloat:startAngle];
         quiverAnim.toValue = [NSNumber numberWithFloat:2 * stopAngle];
         quiverAnim.autoreverses = YES;
         quiverAnim.duration = 0.12;
         quiverAnim.repeatCount = HUGE_VALF;
-        float timeOffset = (float)(arc4random() % 100)/100 - 0.50;
+        CGFloat timeOffset = (arc4random() % 100)/100.0 - 0.50;
         quiverAnim.timeOffset = timeOffset;
         self.quiveringAnimation = quiverAnim;
     }
